requirements.txt
google-cloud-storage
google-cloud-kms
cryptography